+++ src/components/Post.tsx
@@ -24,16 +24,19 @@ function Post({ data, selected }: Props) {
     <Flex
       gap="10px"
       pb="8px"
-      as="button"
-      onClick={handleSelect}
-      disabled={selected}
+      onClick={!selected ? handleSelect : undefined}
       bg={selected ? "gray.200" : "transparent"}
+      cursor={selected ? "default" : "pointer"}
       transition="all 0.2s cubic-bezier(.08,.52,.52,1)"
-      _hover={{ bg: "purple.100" }}
-      _active={{
-        bg: "purple.200",
-        transform: "scale(0.98)",
-      }}
+      _hover={!selected ? { bg: "purple.100" } : undefined}
+      _active={
+        !selected
+          ? {
+              bg: "purple.200",
+              transform: "scale(0.98)",
+            }
+          : undefined
+      }
     >
       {data.thumbnail ? (
         <Image
@@ -48,11 +51,11 @@ function Post({ data, selected }: Props) {
         <Center
           boxSize="140px"
           flexShrink="0"
-          bg="gray.200"
+          bg="gray.300"
           borderRadius="5px"
           mt="8px"
         >
-          <ViewOffIcon boxSize="50px" color="gray.300" />
+          <ViewOffIcon boxSize="50px" color="gray.400" />
         </Center>
       )}
       <Box>