+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## `Reddit Post Client`
 
-### v0.1.2-dev
+### v0.1.3-dev
 
+### v0.1.2
+
+- [ENH] UseFetchPost updated to prevent wrong image urls @christianmotta
+- [NEW] NoImageContainer created to handle non-image posts @christianmotta
+- [NEW] ImageContainer created to handle different image sizes @christianmotta
+- [NEW] EmptyPostDetail created to handle non-selected state @christianmotta
+- [ENH] PostDetail receiving data from the hook @christianmotta
 - [ENH] Selected post behavior added to list @christianmotta
 - [ENH] PostSkeleton created for loading state @christianmotta
 - [NEW] PostsList receiving data from the hook @christianmotta